requirements.txt
appdirs==1.4.4
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
distlib==0.3.1
filelock==3.0.12
idna==2.10
iniconfig==1.1.1
mock==4.0.3
packaging==20.9
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.2
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
toml==0.10.2
urllib3==1.26.3
virtualenv==20.4.2
